+++ src/logic/reference.ts
@@ -8,11 +8,18 @@ function getNodeText(node: Content): string {
   return ''
 }
 
-export function getReferenceText(reference: string, nodes: Content[]): string {
-  const node = nodes.find(
+export function getReferencedNode(
+  reference: string,
+  nodes: Content[]
+): Content | undefined {
+  return nodes.find(
     node =>
       typeof node === 'object' && node && 'id' in node && node.id === reference
   )
+}
+
+export function getReferenceText(reference: string, nodes: Content[]): string {
+  const node = getReferencedNode(reference, nodes)
   if (!node) return ''
   return getNodeText(node)
 }