+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Bug Fixes
 * Fixed `noWrap` in table cell with fixed width
 * Fixed edge cases around table borders
 * Fixed `pageOrientation` if the page width is larger than its height
+* Fixed printing `undefined` as watermark text
 
 Internal