+++ CHANGELOG.md
@@ -8,7 +8,8 @@ Features
 * Added support for complex image definition objects
 * Added support for single-side margins (e.g. `marginLeft`)
 * Added support for `preserveTrailingSpaces`
-* Added support for new `pageBreak` values `...Odd` / `...Even`,
+* Added support for new `pageBreak` values `...Odd` / `...Even`
+* Added support for multiple ToCs
 * Made renderer more resilient to avoid crashes for invalid content
 
 Bug Fixes